Remove duplicate webfinger spec
